chore: fix components type pt. 2
