+++ deno/payloads/v8/channel.ts
@@ -1015,17 +1015,11 @@ export interface APIActionRowComponent extends APIBaseMessageComponent<Component
 	components: Exclude<APIMessageComponent, APIActionRowComponent>[];
 }
 
-export interface ModalSubmitComponent {
-	type: ComponentType;
-	custom_id: string;
-	value: string;
-}
-
 export interface APIModalActionRowComponent extends APIBaseMessageComponent<ComponentType.ActionRow> {
 	/**
 	 * The components in the ActionRow
 	 */
-	components: ModalSubmitComponent[];
+	components: Exclude<APIModalComponent, APIModalActionRowComponent>[];
 }
 
 /**
