+++ deno/payloads/v8/interactions.ts
@@ -16,6 +16,7 @@ export * from './_interactions/messageComponents.ts';
 export * from './_interactions/ping.ts';
 export * from './_interactions/responses.ts';
 export * from './_interactions/applicationCommands.ts';
+export * from './_interactions/modalSubmit.ts';
 
 /**
  * https://discord.com/developers/docs/interactions/receiving-and-responding#interaction-object
